+++ lib/ept_sdk/relationship_empty.ex
@@ -1,0 +1,5 @@
+defmodule EPTSDK.RelationshipEmpty do
+  @enforce_keys [:name]
+  defstruct [:name]
+  # TODO: What if we could turn this into a request to fill the data?
+end